+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Changes
 =======
 
-## 0.1.0
+## UNRELEASED
 
+* Feature: Add `formideploy serve` action.
+  [#8](https://github.com/FormidableLabs/formideploy/issues/8)
 * Initial release.